dcc spec and reorganize controllers to match documentation
