+++ scripts/swagger-generation/package.json
@@ -1,0 +1,7 @@
+{
+  "dependencies": {
+    "@google-cloud/storage": "^5.15.5",
+    "fetch": "^1.1.0",
+    "node-fetch": "^3.1.0"
+  }
+}
